:hammer: :fix: ajuste frotageradores

Approved-by: Wenklesley Mendes Oliveira

+++ backoffice/Domain/Entidades/Cadastros/FrotaGerador/FrotaGerador.cs
@@ -11,9 +11,9 @@ namespace Domain.Entidades.Cadastros.FrotaGerador
         [ForeignKey("ControleDeFrota")]
         public int? IdFrota { get; set; }
         public DateTime? CreatedAt { get; set; }
-        public int? HoraInicio { get; set; }
-        public int? HoraFim { get; set; }
-        public int? HorasUso { get; set; }
+        public double? HoraInicio { get; set; }
+        public double? HoraFim { get; set; }
+        public double? HorasUso { get; set; }
         public DateTime? DataTrocaOleo { get; set; }
         [ForeignKey("Empresa")]
         public int? IdEmpresa { get; set; }
